ng12-risk-assessor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12

google-genai==0.7.0
chromadb==0.5.11
pypdf==5.1.0
requests==2.32.3
numpy==2.1.3
